+++ src/PrettierPackage.cs
@@ -14,21 +14,14 @@ namespace JavaScriptPrettier
     [InstalledProductRegistration("#110", "#112", Vsix.Version, IconResourceID = 400)]
     [Guid(PackageGuids.guidPrettierPackageString)]
     [ProvideMenuResource("Menus.ctmenu", 1)]
-    [ProvideOptionPage(typeof(OptionPageGrid),
-        "Prettier", "General", 0, 0, true)]
+    [ProvideOptionPage(typeof(OptionPageGrid), "Prettier", "General", 0, 0, true)]
     [ProvideAutoLoad(cmdUiContextGuid: VSConstants.UICONTEXT.NotBuildingAndNotDebugging_string, flags: PackageAutoLoadFlags.BackgroundLoad)]
     public sealed class PrettierPackage : AsyncPackage
     {
         internal DTE2 _dte;
-
         internal RunningDocumentTable _runningDocTable;
         internal OptionPageGrid optionPage;
-
         internal ServiceProvider _serviceProvider;
-
-        public PrettierPackage()
-        {
-        }
 
         protected override async System.Threading.Tasks.Task InitializeAsync(CancellationToken cancellationToken, IProgress<ServiceProgressData> progress)
         {
@@ -37,7 +30,6 @@ namespace JavaScriptPrettier
             _dte = ServiceProvider.GlobalProvider.GetService(typeof(EnvDTE.DTE)) as DTE2;
             _serviceProvider = new ServiceProvider((IServiceProvider)_dte);
             _runningDocTable = new RunningDocumentTable(_serviceProvider);
-
             _runningDocTable.Advise(new RunningDocTableEventsHandler(this));
 
             optionPage = (OptionPageGrid)GetDialogPage(typeof(OptionPageGrid));
@@ -49,8 +41,8 @@ namespace JavaScriptPrettier
     public class OptionPageGrid : DialogPage
     {
         [Category("Prettier")]
-        [DisplayName("Run On Save ")]
+        [DisplayName("Format On Save")]
         [Description("Run Pretter whenever a file is saved")]
-        public bool RunOnSave { get; set; }
+        public bool FormatOnSave { get; set; }
     }
 }