+++ src/RunningDocTableEventsHandler.cs
@@ -9,7 +9,7 @@ using Microsoft.VisualStudio.TextManager.Interop;
 
 namespace JavaScriptPrettier
 {
-    class RunningDocTableEventsHandler : IVsRunningDocTableEvents3
+    internal sealed class RunningDocTableEventsHandler : IVsRunningDocTableEvents3
     {
         private readonly PrettierPackage _package;
 
@@ -19,41 +19,35 @@ namespace JavaScriptPrettier
         }
 
         public int OnAfterFirstDocumentLock(uint docCookie, uint dwRDTLockType, uint dwReadLocksRemaining, uint dwEditLocksRemaining) => VSConstants.S_OK;
-
         public int OnBeforeLastDocumentUnlock(uint docCookie, uint dwRDTLockType, uint dwReadLocksRemaining, uint dwEditLocksRemaining) => VSConstants.S_OK;
-
         public int OnAfterSave(uint docCookie) => VSConstants.S_OK;
-
         public int OnAfterAttributeChange(uint docCookie, uint grfAttribs) => VSConstants.S_OK;
-
         public int OnBeforeDocumentWindowShow(uint docCookie, int fFirstShow, IVsWindowFrame pFrame) => VSConstants.S_OK;
-
         public int OnAfterDocumentWindowHide(uint docCookie, IVsWindowFrame pFrame) => VSConstants.S_OK;
-
         public int OnAfterAttributeChangeEx(uint docCookie, uint grfAttribs, IVsHierarchy pHierOld, uint itemidOld, string pszMkDocumentOld, IVsHierarchy pHierNew, uint itemidNew, string pszMkDocumentNew) => VSConstants.S_OK;
 
         public int OnBeforeSave(uint docCookie)
         {
-            if (_package.optionPage.RunOnSave)
+            if (_package.optionPage.FormatOnSave)
             {
-                var docInfo = _package._runningDocTable.GetDocumentInfo(docCookie);
-                var doc = _package._dte.Documents.OfType<Document>().SingleOrDefault(x => x.FullName == docInfo.Moniker);
+                RunningDocumentInfo docInfo = _package._runningDocTable.GetDocumentInfo(docCookie);
+                Document doc = _package._dte.Documents.OfType<Document>().SingleOrDefault(x => x.FullName == docInfo.Moniker);
 
                 if (doc != null)
                 {
-                    var vsTextView = GetIVsTextView(doc.FullName);
+                    IVsTextView vsTextView = GetIVsTextView(doc.FullName);
                     if (vsTextView == null)
                     {
                         return VSConstants.S_OK;
                     }
 
-                    var wpfTextView = GetWpfTextView(vsTextView);
+                    IWpfTextView wpfTextView = GetWpfTextView(vsTextView);
                     if (wpfTextView == null)
                     {
                         return VSConstants.S_OK;
                     }
 
-                    var cmd = wpfTextView.Properties.GetProperty<PrettierCommand>("prettierCommand");
+                    PrettierCommand cmd = wpfTextView.Properties.GetProperty<PrettierCommand>("prettierCommand");
                     if (cmd != null)
                     {
                         ThreadHelper.JoinableTaskFactory.Run(() => cmd.MakePrettierAsync());
@@ -63,22 +57,22 @@ namespace JavaScriptPrettier
             return VSConstants.S_OK;
         }
 
-        IVsTextView GetIVsTextView(string filePath)
+        private IVsTextView GetIVsTextView(string filePath)
         {
             return VsShellUtilities.IsDocumentOpen(_package._serviceProvider, filePath, Guid.Empty, out var uiHierarchy, out uint itemId, out var windowFrame)
                 ? VsShellUtilities.GetTextView(windowFrame) : null;
         }
 
-        static IWpfTextView GetWpfTextView(IVsTextView vTextView)
+        private static IWpfTextView GetWpfTextView(IVsTextView vTextView)
         {
             IWpfTextView view = null;
-            var userData = (IVsUserData)vTextView;
+            IVsUserData userData = (IVsUserData)vTextView;
 
             if (userData != null)
             {
-                var guidViewHost = Microsoft.VisualStudio.Editor.DefGuidList.guidIWpfTextViewHost;
+                Guid guidViewHost = Microsoft.VisualStudio.Editor.DefGuidList.guidIWpfTextViewHost;
                 userData.GetData(ref guidViewHost, out var holder);
-                var viewHost = (IWpfTextViewHost)holder;
+                IWpfTextViewHost viewHost = (IWpfTextViewHost)holder;
                 view = viewHost.TextView;
             }